+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * Fix Cache deleted on HttpTransport exception. (#610) @lucas-zimerman
 * Add SentryScopeStateProcessor #603
 * Add net5.0 TFM to libraries #606
+* Add more logging to CachingTransport #619
+* Bump Microsoft.Bcl.AsyncInterfaces to 5.0.0 #618
 
 ## 3.0.0-alpha.5